third_party/blink/web_tests/fast/events/event-fire-disconnected-bubbling-click-expected.txt

Click events should propagate in disconnected DOM trees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS clickedOnParent is true
PASS clickedOnChild is true
PASS successfullyParsed is true

TEST COMPLETE
